requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coverage==6.2
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.9
django-allauth==0.46.0
django-crispy-forms==1.13.0
django-payments==0.15.0
django-payments-payu==1.2.2
environs==9.3.5
filelock==3.4.2
idna==3.3
iniconfig==1.1.1
marshmallow==3.14.1
oauthlib==3.1.1
packaging==21.3
pipenv==2022.1.8
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
